veloz/requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==5.0.2
django-cors-headers==4.3.1
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
pypng==0.20220715.0
python-barcode==0.15.1
python-dotenv==1.0.1
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
